finish up with the forms
- collect all form data
- format it to be ready for submission to the api endpoint
