+++ package.json
@@ -8,6 +8,7 @@
     "start": "next start"
   },
   "dependencies": {
+    "new-component": "^5.0.2",
     "next": "^12.3.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
