+++ src/components/EventDetailsForm/index.js
@@ -1,0 +1,2 @@
+export * from './EventDetailsForm';
+export { default } from './EventDetailsForm';
