+++ src/components/SessionDetailsForm/index.js
@@ -1,0 +1,2 @@
+export * from './SessionDetailsForm';
+export { default } from './SessionDetailsForm';
