+++ src/pages/newEvent.js
@@ -1,0 +1,60 @@
+import { useState } from 'react';
+import EventDetailsForm from '../components/EventDetailsForm/EventDetailsForm';
+import SessionDetailsForm from '../components/SessionDetailsForm/SessionDetailsForm';
+import { getISODateTime } from '../utils';
+
+export default function NewEvent() {
+  const [formSubmissionProgress, setFormSubmissionProgress] = useState({
+    activeForm: 0,
+    activeSessionSubmission: 0,
+  });
+  const [eventDetails, setEventDetails] = useState({
+    eventName: '',
+    eventLocation: '',
+    eventDate: '',
+    eventStartTime: '',
+    eventEndTime: '',
+  });
+  const [sessionsDetailsArray, setSessionsDetailsArray] = useState([]);
+
+  function handleNewEventSubmit(currentSessionInfo) {
+    let eventInfo = {
+      eventData: {
+        name: eventDetails.eventName,
+        venue: eventDetails.eventLocation,
+        date: getISODateTime(eventDetails.eventDate),
+        startTime: getISODateTime(
+          eventDetails.eventDate,
+          eventDetails.eventStartTime
+        ),
+        endTime: getISODateTime(
+          eventDetails.eventDate,
+          eventDetails.eventEndTime
+        ),
+      },
+      sessionData: [...sessionsDetailsArray, currentSessionInfo],
+    };
+    console.log(eventInfo);
+  }
+
+  return (
+    <>
+      <div className="container max-w-xs mx-auto pt-6">
+        <EventDetailsForm
+          formSubmissionProgress={formSubmissionProgress}
+          setFormSubmissionProgress={setFormSubmissionProgress}
+          eventDetails={eventDetails}
+          setEventDetails={setEventDetails}
+        />
+        <SessionDetailsForm
+          formSubmissionProgress={formSubmissionProgress}
+          setFormSubmissionProgress={setFormSubmissionProgress}
+          sessionsDetailsArray={sessionsDetailsArray}
+          setSessionsDetailsArray={setSessionsDetailsArray}
+          handleNewEventSubmit={handleNewEventSubmit}
+          eventDate={eventDetails.eventDate}
+        />
+      </div>
+    </>
+  );
+}
