+++ src/utils.js
@@ -1,0 +1,4 @@
+export function getISODateTime(date, time = '00:00') {
+  let dateTime = new Date(`${date}T${time}`);
+  return dateTime.toISOString();
+}
