+++ tailwind.config.js
@@ -1,4 +1,7 @@
 module.exports = {
-  content: ['./pages/**/*.{js,ts,jsx,tsx,mdx}'],
+  content: [
+    './src/**/*.{js,ts,jsx,tsx,mdx}',
+  ],
+
   plugins: [require('daisyui')],
 };
